+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 1. [#5716](https://github.com/influxdata/chronograf/pull/5716): Do not fetch tag values when no measurement tags are available.
 1. [#5722](https://github.com/influxdata/chronograf/pull/5722): Filter out roles with unknown organization reference.
 1. [#5724](https://github.com/influxdata/chronograf/pull/5724): Detect actual flux support in flux proxy.
+1. [#5747](https://github.com/influxdata/chronograf/pull/5747): Manage individual execution status per query.
 
 ### Other